refactor: extract plugin watcher module
